+++ src/main/java/com/boardgo/domain/meeting/repository/response/MeetingDetailResponse.java
@@ -22,7 +22,7 @@ public record MeetingDetailResponse(
         String detailAddress,
         Integer limitParticipant,
         MeetingState state,
-        Long shareCount,
+        Integer shareCount,
         Long createMeetingCount,
         // TODO: 찜 여부 추가
         List<String> genres,
